+++ src/wiki/wiki-templates.ts
@@ -2,9 +2,9 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-import indexTemplate from "./index-template.md" with { type: "text" };
-import logTemplate from "./log-template.md" with { type: "text" };
-import schemaTemplate from "./schema-template.md" with { type: "text" };
+import indexTemplate from "../assets/wiki/index-template.md" with { type: "text" };
+import logTemplate from "../assets/wiki/log-template.md" with { type: "text" };
+import schemaTemplate from "../assets/wiki/schema-template.md" with { type: "text" };
 
 export function buildSchemaMd(wikiName: string): string {
   return schemaTemplate.replaceAll("{{WIKI_NAME}}", wikiName);